requirements.txt
# pip install -r requirements.txt

# Base ---------------------------
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# Logging ------------------------
tensorboard>=2.4.1

# Plotting -----------------------
pandas>=1.1.4
seaborn>=0.11.0
 
#Graphics ------------------------
matplotlib>=3.4.3

thop  # FLOPs computation

moviepy
azure.cognitiveservices.vision.computervision
azure.cognitiveservices.vision.face
msrest.authentication
azure.cognitiveservices.speech
azure.ai.textanalytics
azure.core.credentials
